requirements.txt
openai==1.3.8
pyautogen==0.2.2
colorama
pyyaml
tqdm
chromadb
msal
pandas
networkX~=3.2.1
Flask==2.1.3
werkzeug==2.2.2
chainlit==1.0.100
azure-core==1.30.0
azure-identity==1.15.0
azure-kusto-data==4.3.1
azure-kusto-ingest
retrying==1.3.4
numpy==1.24.4